adminPanel/requirements.txt
cryptography==1.7.1
celery==3.1.18
Django==1.10.4
django-allauth==0.30.0
django-braces==1.10.0
django-cors-headers==1.3.1
django-oauth-toolkit==0.10.0
django-vanilla-views==1.0.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
oauthlib==1.0.3
psycopg2==2.6.2
pymongo==3.4.0
python-amazon-product-api==0.2.8
pytz==2016.10
twilio==5.6.0
requests==2.12.4
fuzzywuzzy==0.14.0
pbr==1.10.0
python-Levenshtein==0.12.0
six==1.10.0
stevedore==1.19.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
email==4.0.2
pyjwkest==1.3.2
pandas==0.19.2
redis==2.10.3
XlsxWriter==0.9.6

